+++ index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import { render } from 'react-dom';
-import App from './App.jsx';
+import App from './src/App.jsx';
 
 render(<App />, document.getElementById('contents'));
