+++ package.json
@@ -26,6 +26,7 @@
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
     "react-recorder": "^1.0.0",
+    "react-router-dom": "^4.2.2",
     "webpack-cli": "^2.1.3",
     "webpack-dev-middleware": "^3.1.3",
     "webpack-hot-middleware": "^2.22.2"
