+++ src/sampler-components/Settings.jsx
